feat: add build constraints to remove unnecessary overhead when building for different targets, and to allow seamless integration between them in main
